libraries: reworked download/build structure to reduce redundancy

+++ make.sh
@@ -47,6 +47,9 @@ if [ -f $evallog ]; then
   rm $evallog
 fi
 
+. ./scripts/assert.sh
+. ./scripts/package.sh
+
 #########################
 # general functions
 
@@ -142,66 +145,6 @@ resetstate()
     esac
   fi
 } 
-
-#################################
-# abort on missing files and errors
-
-assertfile()
-{
-  if [ ! -e "$1" ]; then
-    if [ -f "$evallog" ]; then
-      cat "$evallog" | sed '/^$/d'
-    fi
-    echo "ERROR: failed on file $1" 1>&2
-    if [ ! "X$2" = "X" ]; then 
-      echo ">> $2"
-    fi
-    if [ "X$SYS_VERBOSE" = "X" ]; then  
-      echo "BUILD FAILED - configure with verbose option for more information" 
-    else 
-      echo "BUILD FAILED"
-    fi
-    exit 1
-  fi
-}
-
-asserterror()
-{
-  if [ ! $1 = 0 ]; then
-    if [ -f $evallog ]; then
-      cat $evallog | sed '/^$/d'
-    fi
-    echo "ERROR: failed with exit code $1" 1>&2
-    if [ ! "X$2" = "X" ]; then 
-      echo ">> $2"
-    fi
-    if [ "X$SYS_VERBOSE" = "X" ]; then  
-      echo "BUILD FAILED - configure with verbose option for more information" 
-    else 
-      echo "BUILD FAILED"
-    fi
-    exit 1
-  fi
-}
-
-asserttool()
-{
-  for tool in $@; do
-    tool_location=`which $tool 2> /dev/null` 
-    if [ "X$tool_location" = "X" ]; then 
-      echo "ERROR: required tool $tool not found." 
-      echo "Please install a package containing this tool before proceeding."
-    if [ "X$SYS_VERBOSE" = "X" ]; then  
-      echo "BUILD FAILED - configure with verbose option for more information" 
-    else 
-      echo "BUILD FAILED"
-    fi
-      exit 1
-    else 
-      vecho "  => $tool.. ok" 
-    fi
-  done
-}
 
 #################################
 # misc file and directory support
@@ -1249,13 +1192,17 @@ make_library()
   if [ `newerindir $libdir $libfile` = "yes" ]; then
     echo " => $libname.."
     cd $libdir
-    ac_output build.sh
-    quiet=
-    if [ "X$SYS_VERBOSE" = "X" ]; then  
-      quiet="> /dev/null 2> /dev/null"
-    fi
-    veval "$SYS_ENV sh build.sh $quiet"
-    rm build.sh
+    if [ -f make.sh ]; then
+      . ./make.sh
+    else
+      ac_output build.sh
+      quiet=
+      if [ "X$SYS_VERBOSE" = "X" ]; then  
+        quiet="> /dev/null 2> /dev/null"
+      fi
+      veval "$SYS_ENV sh build.sh $quiet"
+      rm build.sh
+    fi 
     cd $here
   fi
 }
